Return all allocated Blocks in the frustum
Updated se::allocator::blocks() to return all currently allocated Blocks
in the frustum, not just the ones that were allocated based on the
values in the depth image. This prevents artifacts in noisy datasets
where one Block might be integrated into in one frame but not in the
next due to a large change in the depth measurement.