Wrap image data in cv::Mat to copy it
cv::Mat::datastart and cv::Mat::dataend seem to be used for ROI
computations so it might not be a good idea to use them this way. Using
a cv::Mat wrapper shouldn't have any impact in performance.

Exposing the underlying vector of se::Image as non-const allows resizing
it without updating the width and height which will can lead to
problems.