Add install target
The library was renamed from SE::Supereight to SRL::Supereight2 to bring
it more in-line with the project name and other SRL projects.

The compile options were changed to apply only to the target so that
they are inherited by projects linking against supereight2.

The minimum CMake version was increased to 3.8 to allow specifying the
C++ standard to use (cxx_std_17) on a per-target basis using
target_compile_features(). This allows users of the library not having
to explicitly specify a C++ standard to use. The earliest supported
Ubuntu version, 18.04, has CMake 3.10 package so this shouldn't be an
issue.