Remove unnecessary inline keywords from visitor
In modern C/C++ inline means that multiple definitions of a function are
allowed, i.e. it can be implemented in a header file. It's not necessary
for templated functions or templated class members since they have to be
implemented in a header. The compiler doesn't need it as a hint for
function inlining.