Update the se::Map dimensions to match the allocated se::Octree
The actually allocated map can be larger than the requested map since
the octree size must be a power of 2. With this change se::Map::getDim()
will return the actual dimensions of the map and not the requested ones.
Reordered the data members so that the octree is initialized first.