Return updated octants from integrateDepth()
The list of updated octants is useful for running various algorithms on
the octree incrementally, e.g. frontier detection or ESDF computation.
The current implementation is a balance between generally useful and
quick to implement so that frontier detection can work. We might want to
rethink the API, especially before adding color and semantics fusion as
it could lead to an explosion of se::MapIntegrator::integrateDepth()
overloads.

Used an overload on the public API (se::MapIntegrator::integrateDepth())
since this is the classic C++ style but a pointer in the internal API
(se::details::IntegrateDepthImpl::integrate()) since it simplifies the
code by having only a single integrate() method for all cases.

Note that the current implementation will return the allocated octants
which are a superset of the updated ones. Returning the actually updated
ones will require modifying the se::Updater::update() functions. It was
left as something to do in a future patch as it doesn't affect the
correctness of frontier computation.