Remove se::Map::setOctree()
It's not used anywhere other than a unit test and it's not very useful.
se::Map is just a thin wrapper around an se::Octree that contains no
extra state so if a new octree is needed a new map should be created.
Modifying the octree of an existing map is also an unnecessary source of
bugs.

This change will also allow changing the octree se::Map member from an
std::shared_ptr to a plain variable.