Remove unnecessary std::shared_ptr from se::Map
The shared_ptr made it easy to write slow code while not providing any
real benefits. No part of the implementation relies on the octree being
a shared_ptr. Removing it also makes the code simpler and easier to
read.

If shared_ptr semantics are necessary in code using supereight2, an
std::shared_ptr<se::Map> should be used instead. If a shared_ptr to an
octree is really necessary, the aliasing shared_ptr constructor can be
used like so (namespaces omitted for brevity):

  shared_ptr<Map> map_ptr(new Map(...));
  shared_ptr<Map::OctreeType> octree_ptr(map_ptr, map_ptr->getOctree());