Fix se::get_field() returning the non-weighted occupancy
In 99% of cases what's needed is the actual field value, i.e. the
weighted occupancy. Using the non-weighted occupancy for interpolation,
gradient computation, meshing, visualizing the field or collision
checking is wrong.

This change will result in field values of greater magnitude in saved
slices and some tiny differences in generated meshes. The field
gradients will also be different but they should still be in the same
direction and with not too different magnitudes.