Fix out of bounds read when accessing scale colours
Added a function to avoid similar issues in the future and to allow
easily changing the behavior for large scale values program-wide.
Defined it in the header so that it can be inlined since it's used in
some performance-critical code.

We don't have any cases in the default configuration that would result
in accessing the colour for a scale greater than 7 but it can happen if
BlockSize is increased. It can also happen when abusing the scale to
store some other data.