Fix gradient artifacts
Gradients are computed at the coarsest scale, up to the maximum block
scale, where the data of all samples is valid. While the scale is
increased the base point from which all sample coordinates are computed
moved. In certain rare cases it could move into a region that is only
allocated at the node-level. In that case no gradient would be computed,
even though the node might have valid data. This issue only affected
occupancy mapping.

It's possible that querying nodes not intended behavior but due to some
bug in the gradient computation. This fix works for now, we can clean-up
the gradient computation and investigate this at a later time.