Remove se::visitor::getInterp() overload
This removes an unused overload while keeping a clean separation between
the single-res and multi-res implementations.

The default argument values allow the multi-res overload to be called
with just 2 arguments. To make this possible the type of the
returned_scale argument had to be changed from int& to int* since
non-const reference arguments can't have default values. The extra
overhead (one nullptr test and one pointer dereference per call) is
negligible compared to the rest of the function.