Remove template from ColourData::Config operator<<
If a template parameter is to the left of the scope resolution operator
(::), C++ won't attempt to deduce it (what's called a non-deduced
context). This is case 1) described here:
https://en.cppreference.com/w/cpp/language/template_argument_deduction#Non-deduced_contexts

The template parameter deduction isn't attempted because there might be
no unique solution. See https://stackoverflow.com/a/25245676 for an
example.

Removing the template parameter from operator<< and just providing
overloads for both possible values, works around the issue by removing
the need for template deduction.