Make Map::Config operator<< a friend
If a template parameter is to the left of the scope resolution operator
(::), C++ won't attempt to deduce it (what's called a non-deduced
context). This is case 1) described here:
https://en.cppreference.com/w/cpp/language/template_argument_deduction#Non-deduced_contexts

The template parameter deduction isn't attempted because there might be
no unique solution. See https://stackoverflow.com/a/25245676 for an
example.

Implementing an overload for each possible template argument combination
in order to make operator<< a non-template function is impractical.
Instead, the only thing that works is making operator<< a friend of
Map::Config AND defining (implementing) it inside the Map::Config class
definition.