Fix free-space propagation in occupancy mapping
The error was that child_ptr->parent() is the same as node_ptr, the
pointer to the node currently being processed, thus it corresponds to
depth and not to depth - 1. This could result in these nodes being
processed later than their parent, depending on their order in the
std::set node_set_[depth - 1], resulting in free-space up-propagation
partially using out-of-date information. Since the nodes placed in the
wrong propagation set had children updated to free, the out-of-date
information could only cause the parent occupancy to be more
conservative. Thus the bug shouldn't have created any safety issues.

In rare cases, the nodes placed in the wrong propagation set would be
freed by their parent. If the parent was processed first in
node_set_[depth - 1], this would cause a use-after-free. This
use-after-free was never detected by the address sanitizer because of
how memory pooling works, that is, the memory of objects freed from the
pool isn't freed from the program, thus remaining valid to access. Due
to the way boost::simple_segregated_storage is implemented (used by
boost::object_pool), the parent pointers of deallocated nodes were
overwritten by pointers to the next freed node. This resulted in the
up-propagation done in freed nodes only affecting freed nodes that were
unreachable from the octree.

All things considered, it looks like the bug didn't have any serious
effects.