Fix updating colour/semantics beyond the far plane
This would cause a segfault for depth values of positive infinity. The
occupancy update works correctly in that case, updating the field to
free, but the back-projection and projection result in garbage values,
causing an out-of-bounds read in the colour and/or semantics image.

This was only an issue in occupancy updates since the TSDF integrators
explicitly test against the far plane.