Use octantops::scale_to_size() where possible
It improves code readability.

Usage candidates found with
    git grep '1 << [^"]'
Not all occurrences involve a scale so some were left intact.

The occurrences in block_impl.hpp weren't modified because they would
result in a circular include. Some octant ops should be moved to a
header that doesn't depend on OctantBase for this to work.